+++ lib/models/places/shared/address_model.dart
@@ -1,16 +1,16 @@
 class Address {
-  final String street;
-  final String city;
-  final String state;
-  final String postalCode;
-  final String countryOrRegion;
+  String? street;
+  String? city;
+  String? state;
+  String? countryOrRegion;
+  String? postalCode;
 
   Address({
-    required this.street,
-    required this.city,
-    required this.state,
-    required this.postalCode,
-    required this.countryOrRegion,
+    this.street,
+    this.city,
+    this.state,
+    this.countryOrRegion,
+    this.postalCode,
   });
 
   factory Address.fromJson(Map<String, dynamic> json) {
@@ -18,8 +18,18 @@ class Address {
       street: json['street'],
       city: json['city'],
       state: json['state'],
-      postalCode: json['postalCode'],
       countryOrRegion: json['countryOrRegion'],
+      postalCode: json['postalCode'],
     );
+  }
+
+  Map<String, dynamic> toJson() {
+    return {
+      'street': street,
+      'city': city,
+      'state': state,
+      'countryOrRegion': countryOrRegion,
+      'postalCode': postalCode,
+    };
   }
 }