+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2
+
+* added example file
+* added toJson Methods to Calendar Models
+* changed createCalendarEvent Method
+* added fetchCalendar Method for own calendar and also for other users calendars
+* added more Calendar related Models
+* added findMeetingTime Method for own Times and also for other users Times
+
 ## 0.0.1
 
 * Initial Open Source release.